+++ packages/eslint/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## next
 
-Support TypeScript
+- Upgrade dependencies to latest major versions; migrate from `@stylistic/eslint-plugin-js` to `@stylistic/eslint-plugin` (rule names change from `@stylistic/js/*` to `@stylistic/*`)
+
+- Support TypeScript
 
 ## v0.1.0